% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_utils.R
\name{categories_graph}
\alias{categories_graph}
\title{Category-based Graph Output}
\usage{
categories_graph(q, p, var_names, esti_Z, esti_Beta, categories)
}
\arguments{
\item{q}{Number of nodes after expansion.}

\item{p}{Number of original variables.}

\item{var_names}{Mapping from nodes to original variables.}

\item{esti_Z}{Estimated context-specific inclusion matrix.}

\item{esti_Beta}{Estimated context-specific Beta matrix.}

\item{categories}{Vector of categories per variable.}
}
\value{
List with adjacency matrices at variable level.
}
\description{
Aggregates context-specific adjacency matrices into variable-level adjacency matrices.
}
