% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_gp_utils.R
\name{compute_Cj_var}
\alias{compute_Cj_var}
\title{Compute Cj Variance and XCXt Matrix}
\usage{
compute_Cj_var(Kernels, Xj, gamma_tilde, w, r)
}
\arguments{
\item{Kernels}{Array of kernel matrices (n x n x K).}

\item{Xj}{Vector of values for variable j.}

\item{gamma_tilde}{Vector of inclusion indicators.}

\item{w}{Vector of kernel weights.}

\item{r}{Precision parameter.}
}
\value{
List with C_j (GP covariance) and XCXt.
}
\description{
Computes covariance and XCXt for one variable's GP component.
}
