% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_utils.R
\name{context_spec_graph}
\alias{context_spec_graph}
\title{Context-specific Graph Output}
\usage{
context_spec_graph(q, post_Bta, post_Z, tags, bfdr)
}
\arguments{
\item{q}{Number of nodes.}

\item{post_Bta}{Posterior Beta matrix.}

\item{post_Z}{Posterior inclusion matrix.}

\item{tags}{Node tags.}

\item{bfdr}{Desired Bayesian FDR.}
}
\value{
List with estimated adjacency matrices.
}
\description{
Processes posterior Beta and inclusion samples into estimated context-specific adjacency matrices.
}
