% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{F_transformation}
\alias{F_transformation}
\title{F Transformation for Predictors}
\usage{
F_transformation(X, type, parameter, cont = FALSE)
}
\arguments{
\item{X}{Predictor matrix or vector.}

\item{type}{Vector of variable types.}

\item{parameter}{Lambda parameter.}

\item{cont}{Logical; if TRUE, transforms continuous variables as well.}
}
\value{
Transformed predictors.
}
\description{
Applies transformation F to predictors depending on type and lambda.
}
