% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{find_lambda}
\alias{find_lambda}
\title{Find Lambda for F Transformation}
\usage{
find_lambda(X, type)
}
\arguments{
\item{X}{Matrix of predictors.}

\item{type}{Vector of variable types.}
}
\value{
Estimated lambda.
}
\description{
Computes the lambda parameter for the F-transformation based on Arkaprava & Dunson (2020).
}
