% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_gp_utils.R
\name{kernel_se}
\alias{kernel_se}
\title{Squared Exponential Kernel}
\usage{
kernel_se(Z, lengthscale = 0.5, sigma_f = 1)
}
\arguments{
\item{Z}{Covariate vector.}

\item{lengthscale}{Lengthscale parameter (default: 0.5).}

\item{sigma_f}{Output scale parameter (default: 1).}
}
\value{
Kernel matrix.
}
\description{
Computes the squared exponential kernel matrix for a given covariate vector.
}
