% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_utils.R
\name{predict_jbmgm}
\alias{predict_jbmgm}
\title{Predict with Bayesian Mixed Graphical Model}
\usage{
predict_jbmgm(
  fit,
  X_new,
  test,
  train,
  Z,
  beta_j_est,
  PPI_Z,
  mcmc_samples = 1000
)
}
\arguments{
\item{fit}{Model fit object from bmgm_gp().}

\item{X_new}{New predictor matrix (for test data).}

\item{test}{Indices of testing data}

\item{train}{Indices of training data}

\item{Z}{Covariate matrix (includes both train and test).}

\item{beta_j_est}{Posterior mean or estimate of beta_j.}

\item{PPI_Z}{Posterior probabilities for kernel components.}

\item{mcmc_samples}{Number of posterior samples to use (default 1000).}
}
\value{
Vector of predicted probabilities (for binary Y).
}
\description{
Generates predictions for new data from a fitted VEL.BMGM model.
}
