% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler_bmgm.R
\name{sampler_bmgm}
\alias{sampler_bmgm}
\title{Gibbs Sampler for Bayesian Mixed Graphical Models}
\usage{
sampler_bmgm(
  n = 1,
  Beta,
  theta,
  type,
  categories,
  lambda,
  M = 1000,
  X_new,
  variables,
  std
)
}
\arguments{
\item{n}{Number of samples to generate.}

\item{Beta}{Precision matrix of the graphical model.}

\item{theta}{A list of parameters for each node.}

\item{type}{Character vector indicating variable types.}

\item{categories}{Vector indicating the number of categories (for categorical variables).}

\item{lambda}{Transformation parameter.}

\item{M}{Number of Gibbs sampler iterations. Default is 1000.}

\item{X_new}{Vector of known values for imputation. Default is `c(0,0,0)`.}

\item{variables}{Indicator vector specifying which variables to update.}

\item{std}{Standard deviation used in the transformation.}
}
\value{
A matrix containing the sampled data.
}
\description{
Implements a Gibbs sampler for Bayesian inference on Mixed Graphical Models.
}
