% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{split_X_cat}
\alias{split_X_cat}
\title{Split Categorical Variables for Design Matrix}
\usage{
split_X_cat(X, type)
}
\arguments{
\item{X}{Matrix of predictors.}

\item{type}{Vector of variable types (e.g. "c", "d", "z", "m").}
}
\value{
List with expanded matrix and number of categories per variable.
}
\description{
Expands categorical variables in X and returns the new design matrix and categories per variable.
}
