% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_update_utils.R
\name{update_beta_j}
\alias{update_beta_j}
\title{Update Beta_j for Bayesian Logistic Model}
\usage{
update_beta_j(X, j, z_y, Cj_var, gamma, omega_pg, beta_0, beta_j_current)
}
\arguments{
\item{X}{Design matrix of predictors.}

\item{j}{Index of predictor to update.}

\item{z_y}{Adjusted response vector.}

\item{Cj_var}{List or array of GP covariance matrices.}

\item{gamma}{Inclusion indicators for predictors.}

\item{omega_pg}{Polya-Gamma weights.}

\item{beta_0}{Intercept.}

\item{beta_j_current}{Current beta_j matrix.}
}
\value{
Updated beta_j vector for predictor j.
}
\description{
Updates the beta_j coefficient vector for a single predictor using Bayesian updates.
}
