% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_update_utils.R
\name{update_tau_w}
\alias{update_tau_w}
\title{Update Tau_w Hyperparameter for GP Priors}
\usage{
update_tau_w(lastgammak_all, lastw_all, a_tau_w, b_tau_w)
}
\arguments{
\item{lastgammak_all}{Inclusion indicators (matrix).}

\item{lastw_all}{GP weights (matrix).}

\item{a_tau_w}{Shape parameter for tau_w prior.}

\item{b_tau_w}{Rate parameter for tau_w prior.}
}
\value{
Updated tau_w value.
}
\description{
Update Tau_w Hyperparameter for GP Priors
}
